Resolve issue #1: link.cmake.ld requires absolute path
[Issue #1](https://github.com/winksaville/baremetal-hi/issues/1)

The primary solution was to create a library (libmycrt0) which contains
startup.S.obj and specify a search path for it on the CMAKE_C_LINK_EXECUTABLE
command line with "-L ${CMAKE_BINARY_DIR}" and a reference to the library
using "-l mycrt0". With that change the ".startup" section becomes:

  .startup  . : { mycrt0:startup.S.obj(.text) }

And thus isn't an absolute path. My first attempt to fix was to somehow
specify a "search path" for "object" files but I don't see any ld options
for that, although it could be there.

I could probably use this technique for both cmake and gnu make, but that
would make the gnu make unnecessarily complex. I wouldn't be surprised that
I do eventually combine them, especially if all of the build systems can
use a "simple" linker script.