gcc not configuring properly on travis-ci.
I've checked in a [gcc-5.2.0](https://github.com/winksaville/gcc-5.2.0)
to my github account and modified its configure script to output copious
amounts of debug.

The problem is that the standard configure script is failing the way I'm
using it. It appears the problem is that on my local system I've got the
gmp, mpfr and mpc libraries installed system wide. But on travis-ci I
don't have them so I wget them. But that's not enough the gcc configure
script is looking for them during configuration and they are not yet
configured. Also, the -I include directive is wrong so I've added
--with-{gmp|mpfr|mpc}-include and executed .configure in each of their
directories. Hopefully that helps.